+++ NPC.gd
@@ -1,6 +1,7 @@
 extends CharacterBody2D
 @onready var timer = $Timer
 @onready var dialogue_box = $"Dialogue Box"
+@onready var animated_sprite_2d = $AnimatedSprite2D
 
 enum {
 	IDLE,
@@ -24,10 +25,11 @@ func _process(delta):
 func move(delta):
 	position =+ dir * speed * delta
 
-func interaction():
+func interaction(a_position:Vector2):
 	dialogue_box._npc_interaction()
 	dialogue_box.visible = true
 	timer.start()
+	animated_sprite_2d.flip_h = true if a_position.x < global_position.x else false
 	
 func choose(array):
 	array.shuffle()