+++ scripts/projectiles/base_projectile.gd
@@ -7,14 +7,18 @@ class_name BaseProjectile
 @export var projectile_force : float
 @export_flags_2d_physics var collison = 0b0000
 @onready var area_2d = $Area2D
+@onready var sprite_2d = $Sprite2D
 
 
 func set_up(pos:Vector2, direction:Vector2):
 	position = pos
 	projectile_force = ((-2.0 * _projectile_resource.distance) / (_projectile_resource.life_time ** 2)) * -1
 	linear_velocity = direction * projectile_force
+	sprite_2d.texture = _projectile_resource.sprite
 	top_level = true
 
+func _process(delta):
+	sprite_2d.rotation = linear_velocity.normalized().angle()
 
 # Called when the node enters the scene tree for the first time.
 func _ready():